refactor: ♻️ removes a few nodes, moved other around
