+++ nodes/fun.py
@@ -1,86 +1,53 @@
 import qrcode
-from ..utils import pil2tensor, here, tensor2np
-from PIL import Image, ImageOps
-from functools import lru_cache
-import numpy as np
-import torch
-import hashlib
-from pathlib import Path
-from ..log import log
-import uuid
-import folder_paths
-import comfy.model_management as model_management
-import subprocess
+from ..utils import pil2tensor
+from PIL import Image
 
+# class MtbExamples:
+#     """MTB Example Images"""
 
-class SaveGif:
-    @classmethod
-    def INPUT_TYPES(cls):
-        return {
-            "required": {
-                "image": ("IMAGE",),
-                "fps": ("INT", {"default": 12, "min": 1, "max": 120}),
-                "resize_by": ("FLOAT", {"default": 1.0, "min": 0.1}),
-                "pingpong": ("BOOL", {"default": False}),
-            }
-        }
+#     def __init__(self):
+#         pass
 
-    RETURN_TYPES = ()
-    OUTPUT_NODE = True
-    CATEGORY = "mtb/IO"
-    FUNCTION = "save_gif"
+#     @classmethod
+#     @lru_cache(maxsize=1)
+#     def get_root(cls):
+#         return here / "examples" / "samples"
 
-    def save_gif(self, image, fps=12, resize_by=1.0, pingpong=False):
-        if image.size(0) == 0:
-            return ("",)
+#     @classmethod
+#     def INPUT_TYPES(cls):
+#         input_dir = cls.get_root()
+#         files = [f.name for f in input_dir.iterdir() if f.is_file()]
+#         return {
+#             "required": {"image": (sorted(files),)},
+#         }
 
-        images = tensor2np(image)
-        images = [frame.astype(np.uint8) for frame in images]
-        if pingpong:
-            reversed_frames = images[::-1]
-            images.extend(reversed_frames)
+#     RETURN_TYPES = ("IMAGE", "MASK")
+#     FUNCTION = "do_mtb_examples"
+#     CATEGORY = "fun"
 
-        height, width, _ = image[0].shape
+#     def do_mtb_examples(self, image, index):
+#         image_path = (self.get_root() / image).as_posix()
 
-        ruuid = uuid.uuid4()
-        out_path = f"{folder_paths.output_directory}/{ruuid}.gif"
+#         i = Image.open(image_path)
+#         i = ImageOps.exif_transpose(i)
+#         image = i.convert("RGB")
+#         image = np.array(image).astype(np.float32) / 255.0
+#         image = torch.from_numpy(image)[None,]
+#         if "A" in i.getbands():
+#             mask = np.array(i.getchannel("A")).astype(np.float32) / 255.0
+#             mask = 1.0 - torch.from_numpy(mask)
+#         else:
+#             mask = torch.zeros((64, 64), dtype=torch.float32, device="cpu")
+#         return (image, mask)
 
-        log.debug(f"Saving a gif file {width}x{height} as {ruuid}.gif")
+#     @classmethod
+#     def IS_CHANGED(cls, image):
+#         image_path = (cls.get_root() / image).as_posix()
 
-        # Prepare the FFmpeg command
-        command = [
-            "ffmpeg",
-            "-y",
-            "-f",
-            "rawvideo",
-            "-vcodec",
-            "rawvideo",
-            "-s",
-            f"{width}x{height}",
-            "-pix_fmt",
-            "rgb24",  # GIF only supports rgb24
-            "-r",
-            str(fps),
-            "-i",
-            "-",
-            "-vf",
-            f"fps={fps},scale={width * resize_by}:-1",  # Set frame rate and resize if necessary
-            "-y",
-            out_path,
-        ]
-
-        process = subprocess.Popen(command, stdin=subprocess.PIPE)
-
-        for frame in images:
-            model_management.throw_exception_if_processing_interrupted()
-            process.stdin.write(frame.tobytes())
-
-        process.stdin.close()
-        process.wait()
-        results = []
-        results.append({"filename": f"{ruuid}.gif", "subfolder": "", "type": "output"})
-        return {"ui": {"images": results}}
-
+#         m = hashlib.sha256()
+#         with open(image_path, "rb") as f:
+#             m.update(f.read())
+#         return m.digest().hex()
 
 
 class QrCode:
@@ -145,5 +112,5 @@ class QrCode:
 
 __nodes__ = [
     QrCode,
-    SaveGif,
+    #  MtbExamples,
 ]
